+++ CHANGELOG.md
@@ -42,6 +42,7 @@ Whitespace conventions:
     * `Hash#fetch_values`
     * `Hash#to_proc`
     * `Struct#dig`
+    * `Kernel#itself`
 - Added safe navigator (`&.`) support. (#1532)
 - Added Random class with seed support. The following methods were reworked to use it:
     * `Kernel.rand`